+++ backend/package.json
@@ -33,12 +33,14 @@
     "cookie-parser": "^1.4.5",
     "multer": "^1.4.3",
     "multer-s3": "^2.10.0",
+    "multer-s3-transform": "^2.10.3",
     "passport": "^0.5.0",
     "passport-jwt": "^4.0.0",
     "passport-naver-v2": "^2.0.8",
     "reflect-metadata": "^0.1.13",
     "rimraf": "^3.0.2",
     "rxjs": "^7.2.0",
+    "sharp": "^0.29.3",
     "swagger-ui-express": "^4.1.6",
     "typeorm-naming-strategies": "^2.0.0"
   },
