+++ backend/src/configs/cors.config.ts
@@ -1,0 +1,4 @@
+export const corsOptions = {
+  origin: process.env.FRONT_LOCALHOST,
+  credentials: true,
+};
