+++ backend/src/configs/index.ts
@@ -1,0 +1,5 @@
+export { setGlobal } from "./global.config";
+export { typeORMConfig } from "./typeorm.config";
+export { multerOption } from "./s3.config";
+export { swaggerOption } from "./swagger.config";
+export { corsOptions } from "./cors.config";
